+++ src/main/java/kitchenpos/application/OrderService.java
@@ -3,9 +3,9 @@ package kitchenpos.application;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import kitchenpos.application.dto.CreateOrderCommand;
-import kitchenpos.application.dto.CreateOrderResponse;
-import kitchenpos.application.dto.common.OrderLineItemCommand;
+import kitchenpos.application.dto.order.CreateOrderCommand;
+import kitchenpos.application.dto.order.CreateOrderResponse;
+import kitchenpos.application.dto.orderlineitem.OrderLineItemCommand;
 import kitchenpos.domain.MenuRepository;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
