+++ src/main/java/kitchenpos/application/dto/menugroup/CreateMenuGroupCommand.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.menugroup;
 
 public class CreateMenuGroupCommand {
