+++ src/main/java/kitchenpos/application/dto/product/CreateProductCommand.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.product;
 
 import java.math.BigDecimal;
