+++ src/main/java/kitchenpos/application/dto/product/ProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto.common;
+package kitchenpos.application.dto.product;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.math.BigDecimal;
