+++ src/main/java/kitchenpos/application/dto/product/SearchProductResponse.java
@@ -1,10 +1,10 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.product;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.math.BigDecimal;
 import kitchenpos.domain.Product;
 
-public class CreateProductResponse {
+public class SearchProductResponse {
 
     @JsonProperty("id")
     private final Long id;
@@ -13,14 +13,14 @@ public class CreateProductResponse {
     @JsonProperty("price")
     private final BigDecimal price;
 
-    private CreateProductResponse(Long id, String name, BigDecimal price) {
+    private SearchProductResponse(Long id, String name, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public static CreateProductResponse from(Product product) {
-        return new CreateProductResponse(product.id(), product.name(), product.price());
+    public static SearchProductResponse from(Product product) {
+        return new SearchProductResponse(product.id(), product.name(), product.price());
     }
 
     public Long id() {
