+++ src/main/java/kitchenpos/application/dto/tablegroup/CreateTableGroupCommand.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.tablegroup;
 
 import java.util.List;
