+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.ui;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.application.MenuGroupService;
-import kitchenpos.application.dto.CreateMenuGroupResponse;
-import kitchenpos.application.dto.SearchMenuGroupResponse;
+import kitchenpos.application.dto.menugroup.CreateMenuGroupResponse;
+import kitchenpos.application.dto.menugroup.SearchMenuGroupResponse;
 import kitchenpos.ui.dto.CreateMenuGroupRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
