+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -7,9 +7,9 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.application.dto.CreateProductCommand;
-import kitchenpos.application.dto.CreateProductResponse;
-import kitchenpos.application.dto.SearchProductResponse;
+import kitchenpos.application.dto.product.CreateProductCommand;
+import kitchenpos.application.dto.product.CreateProductResponse;
+import kitchenpos.application.dto.product.SearchProductResponse;
 import kitchenpos.domain.Price;
 import kitchenpos.domain.Product;
 import org.junit.jupiter.api.Test;
